Refactor agent interaction logic and enhance tool handling
- Supports moving agent polling to the frontend
- Removed the agentHandler.js file and integrated its functionality into agent.js.
- Added constructTools function to manage tool operations and built-in operations more effectively.
- Updated assistants.js to utilize the new agent instructions and tool management.
- Enhanced userDefinedAssistants.js to streamline agent invocation and state management.
- Improved LLM class to support proper stream termination.
