+++ amplify-lambda-js/assistants/agent.js
@@ -1,79 +1,182 @@
-import { DynamoDBClient, QueryCommand } from "@aws-sdk/client-dynamodb";
-import { unmarshall } from "@aws-sdk/util-dynamodb";
 import axios from "axios";
 
 
-export const getLatestAgentState = async function(accessToken, currentUser, sessionId) {
-    const dynamodb = new DynamoDBClient({});
-
-    const params = {
-        TableName: process.env.AGENT_STATE_DYNAMODB_TABLE,
-        KeyConditionExpression: '#user = :user AND #sessionId = :sessionId',
-        ExpressionAttributeNames: {
-            '#user': 'user',
-            '#sessionId': 'sessionId'
-        },
-        ExpressionAttributeValues: {
-            ':user': { S: currentUser },
-            ':sessionId': { S: sessionId }
-        },
-        ScanIndexForward: false,
-        Limit: 1
-    };
-
-    const response = await dynamodb.send(new QueryCommand(params));
-    return response.Items?.[0] ? unmarshall(response.Items[0]) : null;
-}
-
-export const listenForAgentUpdates = async function(accessToken, currentUser, sessionId, onAgentStateUpdate) {
-    let errorsRemaining = 15;
-    while (true) {
-        try {
-            const state = await getLatestAgentState(accessToken, currentUser, sessionId);
-            const shouldContinue = onAgentStateUpdate(state);
-            if (!shouldContinue) {
-                break;
-            }
-            await new Promise(resolve => setTimeout(resolve, 1000));
-        } catch (error) {
-            // It is possible the state has not been written yet, so we will retry a few times
-            console.error("Error checking agent state:", error);
-            if (errorsRemaining <= 0) {
-                break;
-            }
-            else {
-                errorsRemaining--;
-                await new Promise(resolve => setTimeout(resolve, 2000));
-            }
-        }
-    }
-}
-
-
-
 export const invokeAgent = async function(accessToken, sessionId, requestId, prompt, metadata={}) {
-    // Do other async operations here if needed
-    // const someData = await fetchSomeData();
     const endpoint = process.env.AGENT_ENDPOINT;
 
     console.log("Invoking agent with sessionId:", sessionId);
 
-    const response = await axios.post(
-        endpoint,
-        {
-            data: {
-                sessionId,
-                requestId,
-                prompt,
-                metadata
+    try {
+        const response = await axios.post(
+            endpoint,
+            {
+                data: {
+                    sessionId,
+                    requestId,
+                    prompt,
+                    metadata
+                }
+            },
+            {
+                headers: {
+                    Authorization: "Bearer "+accessToken
+                }
             }
-        },
-        {
-            headers: {
-                Authorization: "Bearer "+accessToken
-            }
-        }
-    );
+        );
+        console.log("Agent invocation successful for sessionId:", sessionId);
+        return response.data;
+    } catch (error) {
+        console.error("Failed to invoke agent for sessionId:", sessionId, "Error:", error.message);
+    }
 
-    return response.data;
+}
+
+export const constructTools = (tools) => {
+    if (tools.length === 0) {
+       return {builtInOperations : [],
+               operations : []
+              }
+    }
+
+    const builtIns = tools.filter(tool => {
+        return tool && tool.operation && tool.operation.type && tool.operation.type === "builtIn";
+    })
+
+    // Anything that isn't a builtIn is an op
+    const ops = tools.filter(tool => {
+        return tool && tool.operation && (!tool.operation.type || tool.operation.type !== "builtIn");
+    });
+
+    const makeTool = (tool => {
+        // Filter out empty parameters that are not needed and will
+        // be assigned by AI
+        const filteredParams = Object.fromEntries(
+            Object.entries(tool.parameters || {}).filter(([_, val]) => val.value !== "")
+        );
+
+        let customName = tool.operation.name;
+        if(tool.customName && tool.customName.length > 0) {
+            customName = tool.customName;
+        }
+
+        let customDescription = tool.operation.description;
+        if(tool.customDescription && tool.customDescription.length > 0) {
+            customDescription = tool.customDescription;
+        }
+
+        return {
+            ...tool.operation,
+            name: tool.operation.name,
+            customName: customName,
+            customDescription: customDescription,
+            bindings: filteredParams,
+        }
+    });
+    return {builtInOperations : builtIns.map(makeTool),
+            operations : ops.map(makeTool),
+            }
+}
+
+
+export const agentInstructions = `
+You are an advanced AI assistant with access to specialized functions that enable you to perform tasks beyond conversation. Your primary goal is to help users accomplish their objectives by thoughtfully utilizing these functions when appropriate.
+
+## Core Principles
+
+1. DELIBERATE DECISION-MAKING
+- Stop and think step by step before deciding on an approach
+- Consider multiple solution paths and select the most appropriate one
+- Explicitly reason through trade-offs between different approaches
+- When uncertain, gather more information before proceeding
+
+2. FUNCTION USAGE GUIDELINES
+- Only use functions when necessary to accomplish the user's goal
+- Select the most appropriate function for each task
+- Structure function calls with precise parameters
+- Validate inputs before making function calls
+- Handle errors gracefully and attempt reasonable fallbacks
+
+3. PROBLEM-SOLVING FRAMEWORK
+- Understand: Clarify the user's objective completely before acting
+- Plan: Outline a clear strategy, including which functions to use and in what sequence
+- Execute: Implement the plan methodically, documenting each step
+- Verify: Confirm results match expectations
+- Refine: If outcomes are suboptimal, adjust your approach and try again
+
+4. COMMUNICATION PROTOCOL
+- Explain your reasoning before making function calls
+- Provide clear summaries of function results
+- When presenting multiple options, justify your recommendations
+- Use appropriate technical detail based on user expertise
+
+5. TOOL RESPONSIBILITY
+- Respect rate limits and resource constraints
+- Prioritize user data privacy and security
+- Use the minimal set of functions needed to accomplish the task
+- Acknowledge limitations of available functions
+
+## Function Usage Patterns
+
+For each function type, follow these specialized protocols:
+
+### Data Retrieval Functions
+- Formulate precise queries to minimize irrelevant results
+- Request only necessary data to respect privacy and efficiency
+- Parse and filter responses before presenting to user
+
+### Computation Functions
+- Validate inputs prior to execution
+- Structure complex calculations as smaller, verifiable steps
+- Include appropriate error handling
+
+### External API Functions
+- Format requests according to API documentation
+- Implement appropriate authentication
+- Handle potential network or service failures gracefully
+
+### File Operation Functions
+- Confirm operations on sensitive data
+- Validate file formats before processing
+- Implement safeguards against unintended data loss
+
+## Decision Framework
+When deciding whether to use functions, evaluate:
+1. Is the task impossible to complete conversationally?
+2. Would using a function significantly improve accuracy or efficiency?
+3. Does the user's request implicitly require function capabilities?
+4. Have simpler approaches been exhausted?
+
+## Function Call Format
+When invoking functions:
+1. Use the correct syntax specific to your environment
+2. Include all required parameters
+3. Format parameter values appropriately
+4. Add helpful comments when complexity warrants explanation
+
+## Error Handling Protocol
+
+If you encounter multiple consecutive errors (two or more) while using the same tool or function:
+
+1. Stop immediately and do not make further attempts with that particular tool
+2. Clearly explain to the user:
+- That you've encountered repeated failures with the specific tool
+- A brief, non-technical summary of what you were attempting to accomplish
+- That you're halting further attempts to prevent wasting time or resources
+
+3. Offer alternative approaches when possible:
+- Suggest a different tool or method that might accomplish the same goal
+- Propose breaking down the task into smaller components that might be more manageable
+- Ask if the user has additional information that could help overcome the obstacle
+
+4. Request clear guidance on how to proceed rather than continuing to attempt the same failed approach
+
+Remember that respecting the user's time and providing transparency about limitations is more valuable than persisting with unsuccessful approaches.
+
+
+Remember that your goal is to augment your capabilities through judicious use of functions, not to rely on them when simpler approaches would suffice. Always prioritize user needs and clear communication.`
+
+
+export const getTools = (messages) => {
+    const lastMessage = messages.slice(-1)[0];
+    return lastMessage.configuredTools ?? [];
 }
