+++ amplify-lambda-js/assistants/assistants.js
@@ -16,6 +16,7 @@ import {getDataSourcesByUse, isImage} from "../datasource/datasources.js";
 import {fillInAssistant, getUserDefinedAssistant} from "./userDefinedAssistants.js";
 import { mapReduceAssistant } from "./mapReduceAssistant.js";
 import { ArtifactModeAssistant } from "./ArtifactModeAssistant.js";
+import { agentInstructions, getTools } from "./agent.js"
 
 const logger = getLogger("assistants");
 
@@ -258,24 +259,19 @@ export const getAvailableAssistantsForDataSources = (model, dataSources, assista
     });
 }
 
-const isUserDefinedAssistant = (assistantId) => {
-    return assistantId && assistantId !== "default" && assistantId.startsWith("astp/");
-}
 
 export const chooseAssistantForRequest = async (llm, model, body, dataSources, assistants = defaultAssistants) => {
     logger.info(`Choose Assistant for Request `);
 
     const clientSelectedAssistant = body.options?.assistantId ?? null;
 
-    const lastMessage = body.messages.slice(-1)[0];
-    const hasTools = lastMessage.configuredTools && lastMessage.configuredTools.length > 0;
-
     let selectedAssistant = null;
-    if(clientSelectedAssistant) {
+    if (clientSelectedAssistant) {
         logger.info(`Client Selected Assistant: `, clientSelectedAssistant);
         // For group ast
         const user = llm.params.account.user;
         const token = llm.params.account.accessToken;
+        
         selectedAssistant = await getUserDefinedAssistant(user, defaultAssistant, clientSelectedAssistant, token);
         if (!selectedAssistant) {
             llm.sendStatus(newStatus(
@@ -290,163 +286,23 @@ export const chooseAssistantForRequest = async (llm, model, body, dataSources, a
                 throw new Error("Provided Assistant ID is invalid or user does not have access to this assistant.");
             }
         }
-    }
-    else if(hasTools) {
+    } else if (getTools(body.messages).length > 0) {
         logger.info("Using tools");
-        // For group ast
-        const user = llm.params.account.user;
-
-        const tools = lastMessage.configuredTools;
-
-        const builtIns = tools.filter(tool => {
-            return tool && tool.operation && tool.operation.type && tool.operation.type === "builtIn";
-        })
-
-        // Anything that isn't a builtIn is an op
-        const ops = tools.filter(tool => {
-            return tool && tool.operation && (!tool.operation.type || tool.operation.type !== "builtIn");
-        });
-
-        const makeTool = (tool => {
-            // Filter out empty parameters that are not needed and will
-            // be assigned by AI
-            const filteredParams = Object.fromEntries(
-                Object.entries(tool.parameters || {}).filter(([_, val]) => val.value !== "")
-            );
-
-            let customName = tool.operation.name;
-            if(tool.customName && tool.customName.length > 0) {
-                customName = tool.customName;
-            }
-
-            let customDescription = tool.operation.description;
-            if(tool.customDescription && tool.customDescription.length > 0) {
-                customDescription = tool.customDescription;
-            }
-
-            return {
-                ...tool.operation,
-                name: tool.operation.name,
-                customName: customName,
-                customDescription: customDescription,
-                bindings: filteredParams,
-            }
-        });
-
-        const assistantOps = ops.map(makeTool);
-        const assistantBuiltIns = builtIns.map(makeTool);
-
+        // Note: tools are added in fillInAssistant in order to support tool use for client selected Assistants
+    
         selectedAssistant = fillInAssistant(
             {
                 name: "Amplify Automation",
-                instructions: `
-You are an advanced AI assistant with access to specialized functions that enable you to perform tasks beyond conversation. Your primary goal is to help users accomplish their objectives by thoughtfully utilizing these functions when appropriate.
-
-## Core Principles
-
-1. DELIBERATE DECISION-MAKING
-   - Stop and think step by step before deciding on an approach
-   - Consider multiple solution paths and select the most appropriate one
-   - Explicitly reason through trade-offs between different approaches
-   - When uncertain, gather more information before proceeding
-
-2. FUNCTION USAGE GUIDELINES
-   - Only use functions when necessary to accomplish the user's goal
-   - Select the most appropriate function for each task
-   - Structure function calls with precise parameters
-   - Validate inputs before making function calls
-   - Handle errors gracefully and attempt reasonable fallbacks
-
-3. PROBLEM-SOLVING FRAMEWORK
-   - Understand: Clarify the user's objective completely before acting
-   - Plan: Outline a clear strategy, including which functions to use and in what sequence
-   - Execute: Implement the plan methodically, documenting each step
-   - Verify: Confirm results match expectations
-   - Refine: If outcomes are suboptimal, adjust your approach and try again
-
-4. COMMUNICATION PROTOCOL
-   - Explain your reasoning before making function calls
-   - Provide clear summaries of function results
-   - When presenting multiple options, justify your recommendations
-   - Use appropriate technical detail based on user expertise
-
-5. TOOL RESPONSIBILITY
-   - Respect rate limits and resource constraints
-   - Prioritize user data privacy and security
-   - Use the minimal set of functions needed to accomplish the task
-   - Acknowledge limitations of available functions
-
-## Function Usage Patterns
-
-For each function type, follow these specialized protocols:
-
-### Data Retrieval Functions
-- Formulate precise queries to minimize irrelevant results
-- Request only necessary data to respect privacy and efficiency
-- Parse and filter responses before presenting to user
-
-### Computation Functions
-- Validate inputs prior to execution
-- Structure complex calculations as smaller, verifiable steps
-- Include appropriate error handling
-
-### External API Functions
-- Format requests according to API documentation
-- Implement appropriate authentication
-- Handle potential network or service failures gracefully
-
-### File Operation Functions
-- Confirm operations on sensitive data
-- Validate file formats before processing
-- Implement safeguards against unintended data loss
-
-## Decision Framework
-When deciding whether to use functions, evaluate:
-1. Is the task impossible to complete conversationally?
-2. Would using a function significantly improve accuracy or efficiency?
-3. Does the user's request implicitly require function capabilities?
-4. Have simpler approaches been exhausted?
-
-## Function Call Format
-When invoking functions:
-1. Use the correct syntax specific to your environment
-2. Include all required parameters
-3. Format parameter values appropriately
-4. Add helpful comments when complexity warrants explanation
-
-## Error Handling Protocol
-
-If you encounter multiple consecutive errors (two or more) while using the same tool or function:
-
-1. Stop immediately and do not make further attempts with that particular tool
-2. Clearly explain to the user:
-   - That you've encountered repeated failures with the specific tool
-   - A brief, non-technical summary of what you were attempting to accomplish
-   - That you're halting further attempts to prevent wasting time or resources
-
-3. Offer alternative approaches when possible:
-   - Suggest a different tool or method that might accomplish the same goal
-   - Propose breaking down the task into smaller components that might be more manageable
-   - Ask if the user has additional information that could help overcome the obstacle
-
-4. Request clear guidance on how to proceed rather than continuing to attempt the same failed approach
-
-Remember that respecting the user's time and providing transparency about limitations is more valuable than persisting with unsuccessful approaches.
-
-
-Remember that your goal is to augment your capabilities through judicious use of functions, not to rely on them when simpler approaches would suffice. Always prioritize user needs and clear communication.
-                `,
+                instructions: agentInstructions,
                 description: "Amplify Automation",
                 data: {
                     opsLanguageVersion: "v4",
-                    operations: assistantOps,
-                    builtInOperations: assistantBuiltIns
                 }
             },
             defaultAssistant
         )
-    }
-    else if (body.options.codeInterpreterOnly && (!body.options.api_accessed)) {
+
+    } else if (body.options.codeInterpreterOnly && (!body.options.api_accessed)) {
         selectedAssistant = await codeInterpreterAssistant(defaultAssistant);
         //codeInterpreterAssistant;
     } else if (body.options.artifactsMode && (!body.options.api_accessed)) {
@@ -455,7 +311,7 @@ Remember that your goal is to augment your capabilities through judicious use of
     }
 
     
-    if(selectedAssistant === null) {
+    if (selectedAssistant === null) {
         const status = newStatus({inProgress: true, message: "Choosing an assistant to help..."});
         llm.sendStatus(status);
         llm.forceFlush();
